src/agent_b/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
requests==2.31.0
pypdf2==3.0.1
pyjwt
google-generativeai== 0.5.4
cachetools==5.3.3
python-docx==1.1.2
Pillow==10.3.0
pytesseract==0.3.10